Implemented enough PDO compatibility stuff to start making the mysql driver usable.
Should work:
- Named and question mark parameters
- Columns bindings
- Almost all fetch modes (except for cursors, columns, and lazy mode which falls back to obj)

There are still a few methods missing, and well as attributes management, but it's a good start ;-).

Also, this will need some serious testing before it's finished, because of the lame documentation provided with PDO (not only must this driver 'work', it should mimic the native PDO extension for the features it implements).